Set default dtype to f32 for Rust

+++ m2cgen/interpreters/rust/sigmoid.rs
@@ -1,7 +1,7 @@
-fn sigmoid(x: f64) -> f64 {
-    if x < 0.0_f64 {
-        let z: f64 = x.exp();
-        return z / (1.0_f64 + z);
+fn sigmoid(x: f32) -> f32 {
+    if x < 0.0_f32 {
+        let z: f32 = x.exp();
+        return z / (1.0_f32 + z);
     }
-    1.0_f64 / (1.0_f64 + (-x).exp())
+    1.0_f32 / (1.0_f32 + (-x).exp())
 }
